+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/NonNamespaceOperation.java
@@ -28,6 +28,8 @@ public interface NonNamespaceOperation<T, L, R> extends
   Createable<T>,
   CreateOrReplaceable<T>,
   DryRunable<WritableOperation<T>>,
+  Replaceable<T>,
+  StatusReplaceable<T>,
   Loadable<R> {
 }